Add cross-platform support

+++ FuckMake/src/util/list.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <utility>
 #include <core/core.h>
 
 #define FM_PREALLOC_COUNT 128
@@ -351,7 +352,7 @@ public:
 			if (items[i] == item) return i;
 		}
 
-		return ~0;
+		return (uint64)~0;
 	}
 
 	template<typename K>
@@ -361,7 +362,7 @@ public:
 			if (CmpFunc(items[i], item)) return i;
 		}
 
-		return ~0;
+		return (uint64)~0;
 	}
 
 	inline T& operator[](uint64 index) {
